Address review comments from William
